add shared arg parse
