+++ src/calicost_supervised.py
@@ -14,6 +14,7 @@ import copy
 from pathlib import Path
 import functools
 import subprocess
+from arg_parse import *
 from hmm_NB_BB_phaseswitch import *
 from utils_distribution_fitting import *
 from utils_hmrf import *
@@ -32,292 +33,6 @@ plt.rcParams.update({'font.size': 14})
 
 import mkl
 mkl.set_num_threads(1)
-
-def read_configuration_file(filename):
-    ##### [Default settings] #####
-    config = {
-        "spaceranger_dir" : None,
-        "snp_dir" : None,
-        "output_dir" : None,
-        # supporting files and preprocessing arguments
-        "hgtable_file" : None,
-        "normalidx_file" : None,
-        "tumorprop_file" : None,
-        "supervision_clone_file" : None,
-        "filtergenelist_file" : None,
-        "filterregion_file" : None,
-        "binsize" : 1,
-        "rdrbinsize" : 1,
-        # "secondbinning_min_umi" : 500,
-        "max_nbins" : 1200,
-        "avg_umi_perbinspot" : 1.5,
-        "bafonly" : True,
-        # phase switch probability
-        "nu" : 1,
-        "logphase_shift" : 1,
-        "npart_phasing" : 2,
-        # HMRF configurations
-        "n_clones" : None,
-        "n_clones_rdr" : 2,
-        "min_spots_per_clone" : 100,
-        "min_avgumi_per_clone" : 10,
-        "maxspots_pooling" : 7,
-        "tumorprop_threshold" : 0.5, 
-        "max_iter_outer" : 20,
-        "nodepotential" : "max", # max or weighted_sum
-        "initialization_method" : "rectangle", # rectangle or datadrive
-        "num_hmrf_initialization_start" : 0, 
-        "num_hmrf_initialization_end" : 10,
-        "spatial_weight" : 2.0,
-        "construct_adjacency_method" : "hexagon",
-        "construct_adjacency_w" : 1.0,
-        # HMM configurations
-        "n_states" : None,
-        "params" : None,
-        "t" : None,
-        "t_phaseing" : 1-1e-4,
-        "fix_NB_dispersion" : False,
-        "shared_NB_dispersion" : True,
-        "fix_BB_dispersion" : False,
-        "shared_BB_dispersion" : True,
-        "max_iter" : 30,
-        "tol" : 1e-3,
-        "gmm_random_state" : 0,
-        "np_threshold" : 2.0,
-        "np_eventminlen" : 10,
-        # integer copy number
-        "nonbalance_bafdist" : 1.0,
-        "nondiploid_rdrdist" : 10.0
-    }
-
-    argument_type = {
-        "spaceranger_dir" : "str",
-        "snp_dir" : "str",
-        "output_dir" : "str",
-        # supporting files and preprocessing arguments
-        "hgtable_file" : "str",
-        "normalidx_file" : "str",
-        "tumorprop_file" : "str",
-        "supervision_clone_file" : "str",
-        "filtergenelist_file" : "str",
-        "filterregion_file" : "str",
-        "binsize" : "int",
-        "rdrbinsize" : "int",
-        # "secondbinning_min_umi" : "int",
-        "max_nbins" : "int",
-        "avg_umi_perbinspot" : "float",
-        "bafonly" : "bool",
-        # phase switch probability
-        "nu" : "float",
-        "logphase_shift" : "float",
-        "npart_phasing" : "int",
-        # HMRF configurations
-        "n_clones" : "int",
-        "n_clones_rdr" : "int",
-        "min_spots_per_clone" : "int",
-        "min_avgumi_per_clone" : "int",
-        "maxspots_pooling" : "int",
-        "tumorprop_threshold" : "float", 
-        "max_iter_outer" : "int",
-        "nodepotential" : "str",
-        "initialization_method" : "str",
-        "num_hmrf_initialization_start" : "int", 
-        "num_hmrf_initialization_end" : "int",
-        "spatial_weight" : "float",
-        "construct_adjacency_method" : "str",
-        "construct_adjacency_w" : "float",
-        # HMM configurations
-        "n_states" : "int",
-        "params" : "str",
-        "t" : "eval",
-        "t_phaseing" : "eval",
-        "fix_NB_dispersion" : "bool",
-        "shared_NB_dispersion" : "bool",
-        "fix_BB_dispersion" : "bool",
-        "shared_BB_dispersion" : "bool",
-        "max_iter" : "int",
-        "tol" : "float",
-        "gmm_random_state" : "int",
-        "np_threshold" : "float",
-        "np_eventminlen" : "int",
-        # integer copy number
-        "nonbalance_bafdist" : "float",
-        "nondiploid_rdrdist" : "float"
-    }
-
-    ##### [ read configuration file to update settings ] #####
-    with open(filename, 'r') as fp:
-        for line in fp:
-            if line.strip() == "" or line[0] == "#":
-                continue
-            # strs = [x.replace(" ", "") for x in line.strip().split(":") if x != ""]
-            strs = [x.strip() for x in line.strip().split(":") if x != ""]
-            assert strs[0] in config.keys(), f"{strs[0]} is not a valid configuration parameter! Configuration parameters are: {list(config.keys())}"
-            if strs[1].upper() == "NONE":
-                config[strs[0]] = None
-            elif argument_type[strs[0]] == "str":
-                config[strs[0]] = strs[1]
-            elif argument_type[strs[0]] == "int":
-                config[strs[0]] = int(strs[1])
-            elif argument_type[strs[0]] == "float":
-                config[strs[0]] = float(strs[1])
-            elif argument_type[strs[0]] == "eval":
-                config[strs[0]] = eval(strs[1])
-            elif argument_type[strs[0]] == "bool":
-                config[strs[0]] = (strs[1].upper() == "TRUE")
-            elif argument_type[strs[0]] == "list_str":
-                config[strs[0]] = strs[1].split(" ")
-    # assertions
-    assert not config["spaceranger_dir"] is None, "No spaceranger directory!"
-    assert not config["snp_dir"] is None, "No SNP directory!"
-    assert not config["output_dir"] is None, "No output directory!"
-
-    return config
-
-
-def read_joint_configuration_file(filename):
-    ##### [Default settings] #####
-    config = {
-        "input_filelist" : None,
-        "snp_dir" : None,
-        "output_dir" : None,
-        # supporting files and preprocessing arguments
-        "hgtable_file" : None,
-        "normalidx_file" : None,
-        "tumorprop_file" : None,
-        "supervision_clone_file" : None,
-        "alignment_files" : [],
-        "filtergenelist_file" : None,
-        "filterregion_file" : None,
-        "binsize" : 1,
-        "rdrbinsize" : 1,
-        # "secondbinning_min_umi" : 500,
-        "max_nbins" : 1200,
-        "avg_umi_perbinspot" : 1.5,
-        "bafonly" : True,
-        # phase switch probability
-        "nu" : 1,
-        "logphase_shift" : 1,
-        "npart_phasing" : 2,
-        # HMRF configurations
-        "n_clones" : None,
-        "n_clones_rdr" : 2,
-        "min_spots_per_clone" : 100,
-        "min_avgumi_per_clone" : 10,
-        "maxspots_pooling" : 7,
-        "tumorprop_threshold" : 0.5, 
-        "max_iter_outer" : 20,
-        "nodepotential" : "max", # max or weighted_sum
-        "initialization_method" : "rectangle", # rectangle or datadrive
-        "num_hmrf_initialization_start" : 0, 
-        "num_hmrf_initialization_end" : 10,
-        "spatial_weight" : 2.0,
-        "construct_adjacency_method" : "hexagon",
-        "construct_adjacency_w" : 1.0,
-        # HMM configurations
-        "n_states" : None,
-        "params" : None,
-        "t" : None,
-        "t_phaseing" : 1-1e-4,
-        "fix_NB_dispersion" : False,
-        "shared_NB_dispersion" : True,
-        "fix_BB_dispersion" : False,
-        "shared_BB_dispersion" : True,
-        "max_iter" : 30,
-        "tol" : 1e-3,
-        "gmm_random_state" : 0,
-        "np_threshold" : 2.0,
-        "np_eventminlen" : 10,
-        # integer copy number
-        "nonbalance_bafdist" : 1.0,
-        "nondiploid_rdrdist" : 10.0
-    }
-
-    argument_type = {
-        "input_filelist" : "str",
-        "snp_dir" : "str",
-        "output_dir" : "str",
-        # supporting files and preprocessing arguments
-        "hgtable_file" : "str",
-        "normalidx_file" : "str",
-        "tumorprop_file" : "str",
-        "supervision_clone_file" : "str",
-        "alignment_files" : "list_str",
-        "filtergenelist_file" : "str",
-        "filterregion_file" : "str",
-        "binsize" : "int",
-        "rdrbinsize" : "int",
-        # "secondbinning_min_umi" : "int",
-        "max_nbins" : "int",
-        "avg_umi_perbinspot" : "float",
-        "bafonly" : "bool",
-        # phase switch probability
-        "nu" : "float",
-        "logphase_shift" : "float",
-        "npart_phasing" : "int",
-        # HMRF configurations
-        "n_clones" : "int",
-        "n_clones_rdr" : "int",
-        "min_spots_per_clone" : "int",
-        "min_avgumi_per_clone" : "int",
-        "maxspots_pooling" : "int",
-        "tumorprop_threshold" : "float", 
-        "max_iter_outer" : "int",
-        "nodepotential" : "str",
-        "initialization_method" : "str",
-        "num_hmrf_initialization_start" : "int", 
-        "num_hmrf_initialization_end" : "int",
-        "spatial_weight" : "float",
-        "construct_adjacency_method" : "str",
-        "construct_adjacency_w" : "float",
-        # HMM configurations
-        "n_states" : "int",
-        "params" : "str",
-        "t" : "eval",
-        "t_phaseing" : "eval",
-        "fix_NB_dispersion" : "bool",
-        "shared_NB_dispersion" : "bool",
-        "fix_BB_dispersion" : "bool",
-        "shared_BB_dispersion" : "bool",
-        "max_iter" : "int",
-        "tol" : "float",
-        "gmm_random_state" : "int",
-        "np_threshold" : "float",
-        "np_eventminlen" : "int",
-        # integer copy number
-        "nonbalance_bafdist" : "float",
-        "nondiploid_rdrdist" : "float"
-    }
-
-    ##### [ read configuration file to update settings ] #####
-    with open(filename, 'r') as fp:
-        for line in fp:
-            if line.strip() == "" or line[0] == "#":
-                continue
-            strs = [x.strip() for x in line.strip().split(":") if x != ""]
-            assert strs[0] in config.keys(), f"{strs[0]} is not a valid configuration parameter! Configuration parameters are: {list(config.keys())}"
-            if len(strs) == 1:
-                config[strs[0]] = []
-            elif strs[1].upper() == "NONE":
-                config[strs[0]] = None
-            elif argument_type[strs[0]] == "str":
-                config[strs[0]] = strs[1]
-            elif argument_type[strs[0]] == "int":
-                config[strs[0]] = int(strs[1])
-            elif argument_type[strs[0]] == "float":
-                config[strs[0]] = float(strs[1])
-            elif argument_type[strs[0]] == "eval":
-                config[strs[0]] = eval(strs[1])
-            elif argument_type[strs[0]] == "bool":
-                config[strs[0]] = (strs[1].upper() == "TRUE")
-            elif argument_type[strs[0]] == "list_str":
-                config[strs[0]] = strs[1].split(" ")
-    # assertions
-    assert not config["input_filelist"] is None, "No input file list!"
-    assert not config["snp_dir"] is None, "No SNP directory!"
-    assert not config["output_dir"] is None, "No output directory!"
-
-    return config
 
 
 def main(configuration_file):
